+++ v21_gfp_exact_sched_test/2processors/custom_types/ts.h
@@ -3,19 +3,19 @@
 
 struct TS {
 	// Task system
-    uint_fast8_t n; // n. of tasks -- must be less than MAXN
-	static const uint_fast8_t MAXN = 16;
-    uint_fast8_t C[MAXN]; // WCETs
-    uint_fast8_t D[MAXN]; // deadlines
-    uint_fast8_t P[MAXN]; // minimum interarrival times
-    uint_fast8_t WRTub[MAXN]; // the upper bound for the worst-case response times
+    unsigned short n; // n. of tasks -- must be less than MAXN
+	static const unsigned short MAXN = 16;
+    unsigned short C[MAXN]; // WCETs
+    unsigned short D[MAXN]; // deadlines
+    unsigned short P[MAXN]; // minimum interarrival times
+    unsigned short WRTub[MAXN]; // the upper bound for the worst-case response times
 	
-    uint_fast8_t pmax; // max(P_i) (buffered value)
+    unsigned short pmax; // max(P_i) (buffered value)
 	TS() { }
     
-	TS(uint_fast8_t n_) : n(n_), pmax(0) { assert(n <= MAXN); }
+	TS(unsigned short n_) : n(n_), pmax(0) { assert(n <= MAXN); }
     
-	void setTask(uint_fast8_t i, uint_fast8_t Ci, uint_fast8_t Di, uint_fast8_t Pi) {
+	void setTask(unsigned short i, unsigned short Ci, unsigned short Di, unsigned short Pi) {
 		C[i] = Ci;
 		D[i] = Di;
 		P[i] = Pi;
@@ -26,7 +26,7 @@ struct TS {
 		assert(D[i] <= P[i]); // * constrained deadlines *
 	}
     
-    uint_fast8_t Pmax() const { return pmax; }
+    unsigned short Pmax() const { return pmax; }
 	void read();
 };
